requirements.txt
Django==3.2.9
python-decouple==3.5
psycopg2==2.9.2
django-allauth==0.47.0
django-countries==7.2.1
pyuca==1.2
django-taggit==2.0.0
shortuuid==1.0.8
Pillow==8.4.0
regex==2021.11.10
django-paypal==1.1.2
django-storages==1.12.3
boto3==1.20.26
easy-thumbnails==2.8
opencv-python==4.5.5.62
django-activity-stream==0.10.0